[MISC] Move /{core->utility}/char_operations/predicate_detail.hpp.
